simbioz-landing/server/requirements.txt
python-telegram-bot==12.7
python-dotenv==0.20.0
requests==2.17.0
